Conflicts:
	Core/Component/FPSCameraComponent.cpp
	Core/Component/PhysxComponent.cpp
	Graphics/Camera/FPSCameraController.cpp
	PhysX/PhysXManager.cpp
	Run/Data

+++ Core/Component/PhysxComponent.cpp
@@ -115,11 +115,14 @@ void CPhysxComponent::Init(Vect3f scale, Vect3f position)
 
 void CPhysxComponent::Destroy()
 {
+	if ( m_Destroyed ) return;
+	m_Destroyed = true;
 	CEngine::GetSingleton().getPhysXManager()->destroyActor(getName());
 }
 
 void CPhysxComponent::PhysxUpdate()
 {
+	if ( m_Destroyed ) return;
 	if ( !m_isStatic )
 	{
 		if ( m_isKinematic )
@@ -136,23 +139,24 @@ void CPhysxComponent::PhysxUpdate()
 
 void CPhysxComponent::FixedUpdate(float ElapsedTime)
 {
+	if ( m_Destroyed ) return;
 	auto collisions = CEngine::GetSingleton().getPhysXManager()->getActorCollisions( getName() );
 	auto own = GetOwner();
 	auto cm = CEngine::GetSingleton().getComponentManager();
 	for ( auto &const col : collisions )
-	{		
-		if (cm->get(col) != nullptr) //Can be nullptr if we destroy the object and later check the physx actor colisions
+	{
+		auto other = cm->get( col );
+		if ( other )
 		{
-			auto otherOwner = cm->get(col)->GetOwner();
-			own->SendMsg("OnCollision", otherOwner);			
+			auto otherOwner = other->GetOwner();
+			own->SendMsg( "OnCollision", otherOwner );
 		}
 	}
 }
 
-
-
 void CPhysxComponent::Move(Vect3f position)
 {
+	if ( m_Destroyed ) return;
 	Quatf quat = Quatf::GetQuaternionFromRadians(Vect3f(GetOwner()->GetYaw(), GetOwner()->GetPitch(), GetOwner()->GetRoll()));
 	CEngine::GetSingleton().getPhysXManager()->MoveActor(getName(), position, quat);
 }
